Execute a command

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -4,4 +4,6 @@ exec { 'killmenow_process':
   command     => 'pkill killmenow',
   path        => '/usr/bin:/usr/sbin:/bin:sbin',
   logoutput   => true,
+  refreshonly => true,
+  onlyif      => 'pgrep killmenow',
 }
